Moved private classes into dedicated header
